requirements.txt
Pillow >= 1.1.6